simple: introduce $VESTRICIUS_KEEP_DOWNLOADED
if the environment variable $VESTRICIUS_KEEP_DOWNLOADED is set, the
downloaded files will not be removed.

+++ vestricius/plugins/simple.py
@@ -127,8 +127,10 @@ class SimpleCoreHaruspex(Haruspex):
                               tempfile.gettempdir(),
                               self._on_block_received)
         report = self.inspect(fn)
-        debug(_("Removing '{}'").format(fn))
-        os.unlink(fn)
+        keep = 'VESTRICIUS_KEEP_DOWNLOADED' in os.environ or False
+        if not keep:
+            debug(_("Removing '{}'").format(fn))
+            os.unlink(fn)
         return report
 
     def _on_block_received(self, n_blocks, block_size, n_bytes):
